#8 Comments / Shader Extension
Decided to clear a few comments and improve some of them for more readability, will be keeping half of them so it is easier for me to scroll through it.

Shader Extension has been added which can be found here :
https://marketplace.visualstudio.com/items?itemName=DanielScherzer.GLSL :

Highlights comments within the Shaders.  File extension has been changed to shader.fs to shader.frag , similairy with shader.vs to shader.vertex   Necessary for the specified extension to work.